+++ src/tonpy/tvm/emulator.py
@@ -1,8 +1,11 @@
 # Copyright (c) 2023 Disintar LLP Licensed under the Apache License Version 2.0
 
 from tonpy.libs.python_ton import PyEmulator
-from tonpy.types import VmDict, Cell, CellSlice
-from typing import Union
+
+from tonpy import Stack, StackEntry
+from tonpy.tvm.c7 import BlockId
+from tonpy.types import VmDict, Cell, CellSlice, begin_cell
+from typing import Union, Tuple, List
 
 
 class Emulator:
@@ -22,6 +25,15 @@ class Emulator:
     def emulate_tick_tock_transaction(self, shard_account: Cell, is_tock: bool, unixtime: int, lt: int) -> bool:
         return self.emulator.emulate_tick_tock_transaction(shard_account.cell, is_tock, str(unixtime), str(lt),
                                                            1 if lt >= 3709412000000 else 0)
+
+    def set_prev_blocks_info(self, prev_blocks_info: Union[Tuple[List[BlockId], BlockId], Tuple[List, List]]):
+        if len(prev_blocks_info) > 0 and len(prev_blocks_info[0]) > 0 and isinstance(prev_blocks_info[0][0], BlockId):
+            prev_blocks_info[0] = [i.to_data() for i in prev_blocks_info[0]]
+
+        if isinstance(prev_blocks_info[1], BlockId):
+            prev_blocks_info[1] = prev_blocks_info[1].to_data()
+
+        self.emulator.set_prev_blocks_info(StackEntry(value=prev_blocks_info).entry)
 
     def set_rand_seed(self, seed: Union[int, hex]) -> None:
         if isinstance(seed, int):
@@ -68,3 +80,11 @@ class Emulator:
             return c.begin_parse()
         else:
             return c
+
+
+if __name__ == "__main__":
+    cfg = VmDict(32)
+    cfg.set_builder(0, begin_cell())
+
+    em = Emulator(cfg)
+    em.set_prev_blocks_info([[0], 0])